requirements.txt
streamlit
ultralytics
opencv-python-headless
pillow
matplotlib
pandas
seaborn
